AiWorker/requirements.txt
redis
psycopg2-binary
google-generativeai
python-dotenv